==> notebooks/01_json_scraping.ipynb <==
## JSON modul importálása

import json

## JSON fájl beolvasása

# json fájl útvonala
filepath = r'/data/json_scraping/spotify_playlist.json'

with open(filepath, 'r') as file:
	data = json.load(file) # data tartalma a JSON tartalma

## számok adatainak kiírása

# 0. zene címe
data['contents']['items'][0]['name']

# minden előadó neve
for track in data['contents']['items']:
	for artist in track['artists']:
		print(artist['name'])
	print()

# melyik számot hányszor játszották le
for track in data['contents']['items']:
	print(track['name'], ',', track['playCount'])
