==> notebooks/03_get_taxi_data.ipynb <==
from datetime import datetime
from dateutil.relativedelta import relativedelta
import requests

### egy trip

url = "https://data.cityofchicago.org/resource/ajtu-isnz.json?trip_id=000001dd3888f74eaf131b0627d367e8c5b73cc4"
response = requests.get(url)
data = response.json()
data

### egy nap tripjei

# éjféltől éjfélig, max 30 000 sor
url = (f"https://data.cityofchicago.org/resource/ajtu-isnz.json?"
	   f"$where=trip_start_timestamp >= '2024-02-16T00:00:00' "
	   f"AND trip_start_timestamp <= '2024-02-16T23:59:59'"
	   f"&$limit=30000")
response = requests.get(url)
data = response.json()
len(data)

### egy kiválasztott nap tripjei

# mai dátumhoz képest 2 hónappal korábban
current_datetime = datetime.now() - relativedelta(months = 2) # datetime típus
formatted_datetime = current_datetime.strftime("%Y-%m-%d") # string from time formázás
formatted_datetime

# éjféltől éjfélig, max 30 000 sor
url = (f"https://data.cityofchicago.org/resource/ajtu-isnz.json?"
	   f"$where=trip_start_timestamp >= '{formatted_datetime}T00:00:00' "
	   f"AND trip_start_timestamp <= '{formatted_datetime}T23:59:59'"
	   f"&$limit=30000")
response = requests.get(url)
data = response.json()
len(data)
