==> notebooks/04_get_weather_data.ipynb <==
from datetime import datetime
from dateutil.relativedelta import relativedelta

import pandas as pd
import requests

current_datetime = datetime.now() - relativedelta(months = 2) # datetime típus
formatted_datetime = current_datetime.strftime("%Y-%m-%d") # string from time formázás
url = "https://archive-api.open-meteo.com/v1/era5"
params = {
	'latitude': 41.85, # chicago
	'longitude': -87.65,
	'start_date': formatted_datetime,
	'end_date': formatted_datetime,
	'hourly': 'temperature_2m,wind_speed_10m,rain,precipitation'
}
response = requests.get(url, params = params)
data = response.json()

# api lekérdezés adatai metaadatok nélkül
weather_data = {
	'datetime':data['hourly']['time'],
	'temperature':data['hourly']['temperature_2m'],
	'wind_speed':data['hourly']['wind_speed_10m'],
	'rain':data['hourly']['rain'],
	'precipitation':data['hourly']['precipitation']}

weather_df = pd.DataFrame(weather_data)
weather_df['datetime'] = pd.to_datetime(weather_df['datetime'])

weather_df.head()

weather_df.info()
