==> notebooks/06_taxi_data_mapping.ipynb <==
### EDA: exploratory data analysis

from datetime import datetime
from dateutil.relativedelta import relativedelta

import pandas as pd
pd.set_option('display.max_columns', 30)

import requests

current_datetime = datetime.now() - relativedelta(months = 2)
formatted_datetime = current_datetime.strftime("%Y-%m-%d")

url = (f"https://data.cityofchicago.org/resource/ajtu-isnz.json?"
	   f"$where=trip_start_timestamp >= '{formatted_datetime}T00:00:00' "
	   f"AND trip_start_timestamp <= '{formatted_datetime}T23:59:59'"
	   f"&$limit=30000")
response = requests.get(url)
data = response.json()
taxi_trips = pd.DataFrame(data)

## adattranszformáció, időjárás és dátum dimenziók segédoszlopai

# NaN értékek kezelése
taxi_trips = taxi_trips.drop(['pickup_census_tract', 'dropoff_census_tract'], axis = 1)
taxi_trips = taxi_trips.drop(['pickup_centroid_location', 'dropoff_centroid_location'], axis = 1)

# eldobjuk azokat a sorokat, amelyekben null van
taxi_trips.dropna(inplace = True)

# oszlopok átnevezése
taxi_trips = taxi_trips.rename(
		columns = {'pickup_community_area' : 'pickup_community_area_id', 'dropoff_community_area' : 'dropoff_community_area_id'})

# időjárás adatok segéd oszlop - adott trip start timestamp órára lekerekített értéke
taxi_trips['trip_start_timestamp'] = pd.to_datetime(taxi_trips['trip_start_timestamp'])
taxi_trips['datetime_for_weather'] = taxi_trips['trip_start_timestamp'].dt.floor('h')  # órára lefelé kerekít

taxi_trips['trip_start_date'] = pd.to_datetime(taxi_trips['trip_start_timestamp']).dt.date

## időjárás adatok összefűzése

current_datetime = datetime.now() - relativedelta(months = 2)  # datetime típus
formatted_datetime = current_datetime.strftime("%Y-%m-%d")  # string from time formázás

url = "https://archive-api.open-meteo.com/v1/era5"

params = {
	'latitude' : 41.85,  # chicago
	'longitude' : -87.65,
	'start_date' : formatted_datetime,
	'end_date' : formatted_datetime,
	'hourly' : 'temperature_2m,wind_speed_10m,rain,precipitation'
}

response = requests.get(url, params = params)
data = response.json()

# api lekérdezés adatai metaadatok nélkül
weather_data = {
	'datetime' : data['hourly']['time'],
	'temperature' : data['hourly']['temperature_2m'],
	'wind_speed' : data['hourly']['wind_speed_10m'],
	'rain' : data['hourly']['rain'],
	'precipitation' : data['hourly']['precipitation']}

weather_df = pd.DataFrame(weather_data)
weather_df['datetime'] = pd.to_datetime(weather_df['datetime'])

taxi_trips_weather = taxi_trips.merge(weather_df, left_on = 'datetime_for_weather', right_on = 'datetime')

## adattípus konverziók

# eredeti verzió tárolása
taxi_trips_original = taxi_trips.copy()

data_types = {
	"trip_end_timestamp" : "datetime64[ns]",
	"trip_seconds" : "int32",
	"trip_miles" : "float",
	"pickup_community_area_id" : "int8",  # -127/+127
	"dropoff_community_area_id" : "int8",
	"fare" : "float",
	"tips" : "float",
	"tolls" : "float",
	"extras" : "float",
	"trip_total" : "float"
}

taxi_trips = taxi_trips.astype(data_types)

taxi_trips.describe()

#### memóriahasználat

# összeadja, mennyi memóriát használ a df
orig_mem = taxi_trips_original.memory_usage(deep = True).sum()
opt_mem = taxi_trips.memory_usage(deep = True).sum()

reduction = (orig_mem - opt_mem) / orig_mem * 100

print(f'Original memory usage: {orig_mem} bytes\n'
	  f'Optimized memory usage: {opt_mem} bytes\n'
	  f'Memory reduced by: {reduction} %')

## sanity check

# legutolsó trip
taxi_trips[taxi_trips['trip_end_timestamp'] == taxi_trips['trip_end_timestamp'].max()]

# leghosszabb trip
taxi_trips[taxi_trips['trip_seconds'] == taxi_trips['trip_seconds'].max()]

# legdrágább út
taxi_trips[taxi_trips['fare'] == taxi_trips['fare'].max()]

## Dimenziótáblák létrehozása

### payment type dimension

dim_payment_type = taxi_trips['payment_type'].drop_duplicates().reset_index(drop = True)

dim_payment_type = pd.DataFrame({
	'payment_type_id' : range(1, len(dim_payment_type) + 1),
	'payment_type' : dim_payment_type
})

### company dimension

dim_company = taxi_trips['company'].drop_duplicates().reset_index(drop = True)

dim_company = pd.DataFrame({
	'company_id' : range(1, len(dim_company) + 1),
	'company' : dim_company
})

## visszahelyettesítés az eredeti adatokba

fact_taxi_trips = taxi_trips.merge(dim_payment_type, on = 'payment_type', how = 'left')
fact_taxi_trips = fact_taxi_trips.merge(dim_company, on = 'company', how = 'left')

fact_taxi_trips.drop(['payment_type', 'company'], axis = 1, inplace = True)

# dim_payment_type.to_csv('dim_payment_type.csv', index = False)
# dim_company.to_csv('dim_company.csv', index = False)

## Update logika a payment dimenziótábláknak

# új adat szimulálása
dummy_payment_type_data = [{'payment_type' : 'Credit Card'},
						   {'payment_type' : 'Z'},
						   {'payment_type' : 'Y'},
						   {'payment_type' : 'Z'}]
dummy_payment_type_data_df = pd.DataFrame(dummy_payment_type_data)
# ismétlődések eltávolítása
todays_payment_types = pd.DataFrame(dummy_payment_type_data_df['payment_type'].unique(), columns = ['payment_type'])

# új típusok, amik nincsenek benne a dimben
new_payment_types = todays_payment_types[~todays_payment_types['payment_type'].isin(dim_payment_type['payment_type'])]

if not new_payment_types.empty :
	max_id = dim_payment_type['payment_type_id'].max()
	new_payment_types['payment_type_id'] = range(max_id + 1, max_id + 1 + len(new_payment_types))
	dim_payment_type = pd.concat([dim_payment_type, new_payment_types], ignore_index = True)

# új adat szimulálása
dummy_payment_type_data = [{'payment_type' : 'Credit Card'},
						   {'payment_type' : 'X'},
						   {'payment_type' : 'Y'},
						   {'payment_type' : 'X'}]
dummy_payment_type_data_df = pd.DataFrame(dummy_payment_type_data)
# ismétlődések eltávolítása
todays_payment_types = pd.DataFrame(dummy_payment_type_data_df['payment_type'].unique(), columns = ['payment_type'])

# új típusok, amik nincsenek benne a dimben
new_payment_types = todays_payment_types[~todays_payment_types['payment_type'].isin(dim_payment_type['payment_type'])]

if not new_payment_types.empty :
	max_id = dim_payment_type['payment_type_id'].max()
	new_payment_types['payment_type_id'] = range(max_id + 1, max_id + 1 + len(new_payment_types))
	dim_payment_type = pd.concat([dim_payment_type, new_payment_types], ignore_index = True)

## Company dimenziótábla update logika

# új adat szimulálása
dummy_company_data = [{'company' : 'Metro Jet Taxi A.'},
					  {'company' : 'X'},
					  {'company' : 'Y'},
					  {'company' : 'X'}]
dummy_company_df = pd.DataFrame(dummy_company_data)
# ismétlődések eltávolítása
todays_companies = pd.DataFrame(dummy_company_df['company'].unique(), columns = ['company'])

# új típusok, amik nincsenek benne a dimben
new_companies = todays_companies[~todays_companies['company'].isin(dim_company['company'])]

if not new_companies.empty :
	max_id = dim_company['company_id'].max()
	new_companies['company_id'] = range(max_id + 1, max_id + 1 + len(new_companies))
	dim_company = pd.concat([dim_company, new_companies], ignore_index = True)
